Startet moveset crawler implementation and Pymap fixes.

+++ include/rtc.h
@@ -31,7 +31,7 @@ u8 rtc_read_byte();
 u16 special_get_rtc();
 void rtc_chip_wait();
 u8 to_dec(u8 value);
-int rtc_timestamp_to_seconds(rtc_timestamp *t);
+u64 rtc_timestamp_to_seconds(rtc_timestamp *t);
 bool rtc_test();
 
 #endif
